pyproject.toml
[tool.poetry]
name = "authLogin"
version = "0.1.0"
description = "authLogin with sqlalchemy base on pydantic"
authors = ["bagus <bagusssyamsu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
Jinja2 = "^3.1.2"
SQLAlchemy = "^1.4.37"
python-dotenv = "^0.20.0"
bcrypt = "^3.2.2"
fastapi-jwt-auth = "^0.5.0"
itsdangerous = "^2.1.2"
mysqlclient = "^2.1.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
start = "authlogin.app:start"
test = "tests.test_authlogin:test_version"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
